GH-1480: Polishing
