--- master/front-end/src/layout/AdminHeader/Help/HelpLicenseItem/LicenseModal/LicenseDetails/LicenseDetailsFooter/styles.module.css
@@ -1,4 +1,0 @@
-.ActionsWrapper {
-    display: flex;
-    column-gap: 8px;
-}
