+++ master/front-end/src/layout/AdminHeader/Help/HelpLicenseItem/LicenseModal/OSSLicenseDetails/styles.module.css
@@ -1,0 +1,36 @@
+.ContentWrapper {
+    display: flex;
+    flex-direction: column;
+    row-gap: 20px;
+}
+
+.Meta {
+    display: flex;
+    align-items: center;
+    column-gap: 20px;
+    margin-top: 10px;
+}
+
+.MetaStatus {
+    display: flex;
+    align-items: center;
+    column-gap: 8px;
+}
+
+.MetaStatusOSS {
+    color: var(--axelix-bold-grey);
+}
+
+.LicenseLink {
+    color: var(--axelix-primary-blue);
+    transition: color 0.3s;
+
+    &:hover {
+        color: var(--axelix-primary-light-blue);
+    }
+}
+
+.ActionsWrapper {
+    display: flex;
+    column-gap: 8px;
+}
